+++ src/config/mod.rs
@@ -4,5 +4,6 @@ mod config;
 mod test_config;
 
 pub use config::{
-    config_path, read_config, serialize_config, write_config, Config, RepositoryConfig,
+    config_path, deserialize_config, read_config, serialize_config, write_config, Config,
+    RepositoryConfig, RulesConfig,
 };